+++ src/main.cpp
@@ -37,7 +37,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.7.1"};
+constexpr auto version = std::string_view{"0.8.0"};
 
 struct BootstrapConfigResult final
 {